backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
environs
django-filter
Pillow
reportlab
twilio
django-unfold